Split everything up into separate manageable files
