+++ package.json
@@ -11,6 +11,7 @@
     "react-dom": "^16.13.1",
     "react-lazy-hero": "^2.0.0",
     "react-parallax": "^3.0.3",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
     "styled-components": "^5.0.1"
   },
