+++ src/components/Contact.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import {Container} from 'react-bootstrap';
+
+export class Contact extends React.Component {
+    render() {
+      return (
+          <Container>
+              <h1>Contact Us</h1>
+          </Container>
+      );
+    }
+}
